% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabModelEstimates.R
\name{grabModelEstimates}
\alias{grabModelEstimates}
\title{Wrangle important information from fitted models}
\usage{
grabModelEstimates(fitted_model_clogit, fitted_model_glmm)
}
\arguments{
\item{fitted_model_clogit}{A fitted conditional logistic model object, from
\code{\link[survival]{clogit}}.}

\item{fitted_model_glmm}{A fitted generalized linear (logistic) mixed model
object, from \code{\link[lme4]{glmer}}.}
}
\value{
A list including the estimated components from the
\code{\link[survival]{clogit}} model object, the estimated coefficients from
the \code{\link[lme4]{glmer}} model object, and a vector of indices to match
the coefficient names between the models.
}
\description{
This function selects the estimated coefficients from each model, and the
estimated variance component from the clogit model object. It then attempts
to match the coefficients, if possible, and throws errors if not.
}
\details{
It is important to use the same coefficient names in the model
formulas for both models. It is required that the coefficients names in the
clogit model are a subset of those from the glmm. Future versions may have
improved matching functionality without this requirement.
}
\examples{

library(glmerGOF)
set.seed(1)
n <- 50
m <- 4
beta <- c(2, 2)
id <- rep(1:n, each=m)
x <- rnorm(m*n)
b <- rep(rnorm(n), each=m)
y <- rbinom(m*n, 1, plogis(cbind(1, x) \%*\% beta + b))
my_data <- data.frame(y,x,id)

variable_names <- list(DV = "y", grouping = "id")

library(lme4)
fit_glmm <- lme4::glmer(
  formula = y ~ x + (1|id),
  family = "binomial",
  data = my_data
)
library(survival)
fit_clogit <- survival::clogit(
  formula = y ~ x + strata(id),
  data = my_data,
  method = "exact"
)

model_ests <- grabModelEstimates(
  fitted_model_clogit = fit_clogit,
  fitted_model_glmm = fit_glmm
)
names(model_ests)

}
