% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary_TCGOF}
\alias{print.summary_TCGOF}
\title{Prints a "summary_TCGOF" object}
\usage{
\method{print}{summary_TCGOF}(x, ...)
}
\arguments{
\item{x}{object of class "summary_TCGOF"}

\item{...}{dots}
}
\description{
Prints a "summary_TCGOF" object
}
\author{
Brian G. Barkley
}
