% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.TCGOF}
\alias{print.TCGOF}
\title{Prints a "TCGOF" object}
\usage{
\method{print}{TCGOF}(x, ...)
}
\arguments{
\item{x}{object of class "TCGOF"}

\item{...}{dots}
}
\description{
Prints a "TCGOF" object
}
\author{
Brian G. Barkley
}
