% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.TCGOF}
\alias{summary.TCGOF}
\title{Summarizes a "TCGOF" Object}
\usage{
\method{summary}{TCGOF}(object, ...)
}
\arguments{
\item{object}{object of class "TCGOF"}

\item{...}{dots}
}
\value{
\describe{
\item{results}{The results of the test}
\item{test_components}{Inputs for \code{\link{computeD}}.}
\item{glmm_components}{Components from the glmm model.}
\item{clogit_components}{Components from the clogit model.}
\item{prepared_data}{Modeled data after rows with \code{NA}s have been
removed.}
\item{var_names}{Important variable names.}
}
}
\description{
Summarizes a "TCGOF" Object
}
\author{
Brian G. Barkley
}
