% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGOF.R
\name{testGOF}
\alias{testGOF}
\title{Test Goodness Of Fit of Gaussian Random Effect Distribution in Logistic
Mixed Model}
\usage{
testGOF(data, fitted_model_clogit, fitted_model_glmm, var_names = list(DV
  = NULL, grouping = NULL), gradient_derivative_method = NULL,
  use_Richardson_derivative = NULL)
}
\arguments{
\item{data}{The original dataset.}

\item{fitted_model_clogit}{A fitted conditional logistic model object, from
\code{\link[survival]{clogit}}.}

\item{fitted_model_glmm}{A fitted generalized linear (logistic) mixed model
object, from \code{\link[lme4]{glmer}}.}

\item{var_names}{A list indicating the column names (as strings) of two
important variables}

\item{gradient_derivative_method}{User must select either 'simple' (faster)
or 'Richardson' (slower but more accurate) as the desired method for the
numerical derivative procedure. For more information, see the \code{method}
argument in \code{\link[numDeriv]{grad}}.}

\item{use_Richardson_derivative}{defaults to FALSE}
}
\value{
An object of class \code{"TCGOF"}, with the following list items:
\describe{
\item{results}{The test statistic and p-value. See the \code{print()}
generic.}
\item{var_names}{The list of \code{var_names}.}
\item{split_data}{A list object of the data after it has been split into
clusters and NA rows have been removed.}
\item{fits}{Estimated components from the two models, and other important
computations. See the \code{summary()} generic.}
}
}
\description{
This function implements the GOF test introduced Tchetgen & Coull (2006)
Biometrika. The test is designed to determine whether the presumed Gaussian
distribution of the random effect in a logsitic mixed model has adequate fit.
A major contribution of the glmerGOF package is to estimate the variance
matrix from a \code{\link[lme4]{glmer}} model for computing the chi-squared
statistic introduced in Tchetgen & Coull (2006).
}
\details{
The method is based on an asymptotically chi-squared test statistic
comparing the difference between estimated components of a generalized linear
mixed model (\emph{glmm}) and a conditional logistic model (\emph{clogit}).
Before using this function, the user should fit his/her desired glmm with
\code{\link[lme4]{glmer}}, and a clogit model with
\code{\link[survival]{clogit}}. The fitted model objects are required
arguments, and their estimated components are wrangled in
\code{\link{grabModelEstimates}}.
}
\examples{

library(glmerGOF)
set.seed(1)
n <- 50
m <- 4
beta <- c(2, 2)
id <- rep(1:n, each=m)
x <- rnorm(m*n)
b <- rep(rnorm(n), each=m)
y <- rbinom(m*n, 1, plogis(cbind(1, x) \%*\% beta + b))
my_data <- data.frame(y,x,id)

variable_names <- list(DV = "y", grouping = "id")

library(lme4)
fit_glmm <- lme4::glmer(
  formula = y ~ x + (1|id),
  family = "binomial",
  data = my_data
)
library(survival)
fit_clogit <- survival::clogit(
  formula = y ~ x + strata(id),
  data = my_data,
  method = "exact"
)

test_results <- testGOF(
  data = my_data,
  fitted_model_clogit = fit_clogit,
  fitted_model_glmm  = fit_glmm,
  var_names = variable_names,
  gradient_derivative_method = "simple"
)
test_results





}
\references{
Tchetgen Tchetgen, E. J., & Coull, B. A. (2006) \emph{A Diagnostic Test for the Mixing
Distribution in a Generalised Linear Mixed Model.} Biometrika, 93(4),
1003-1010. \url{https://doi.org/10.1093/biomet/93.4.1003}
}
